task01/docker/requirements.txt
kopf==1.37.5
kubernetes==32.0.1
kafka-python==2.1.5
